model/requirements.txt
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
bcrypt==4.1.2
bidict==0.23.1
blinker==1.6.2
CacheControl==0.14.0
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
contourpy==1.2.1
cryptography==41.0.7
cycler==0.12.1
distlib==0.3.7
distro==1.9.0
Django==5.0.6
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
docstring_parser==0.16
docutils==0.20.1
env==0.1.0
executing==2.0.1
Faker==22.4.0
filelock==3.12.4
firebase-admin==6.5.0
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Caching==2.3.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
ggshield==1.24.0
google-ai-generativelanguage==0.6.6
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-cloud-aiplatform==1.57.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.7.1
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
greenlet==3.0.2
grpc-google-iam-v1==0.13.1
grpcio==1.64.1
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
icecream==2.1.3
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.2
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
Kivy-Garden==0.1.5
kiwisolver==1.4.5
klib==1.1.2
Mako==1.3.0
MarkupSafe==2.1.3
marshmallow==3.18.0
marshmallow-dataclass==8.5.14
matplotlib==3.8.4
msgpack==1.0.8
mypy-extensions==1.0.0
numpy==1.26.4
oauthlib==3.2.2
openai==1.36.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
platformdirs==3.0.0
plotly==5.20.0
proto-plus==1.24.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
pygitguardian==1.13.0
Pygments==2.18.0
PyJWT==2.6.0
pymongo==4.7.3
pyparsing==3.1.2
pypiwin32==223
python-dateutil==2.8.2
python-dotenv==0.21.1
python-engineio==4.9.0
python-socketio==5.11.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
rich==12.5.1
rsa==4.9
scipy==1.13.0
screeninfo==0.8.1
seaborn==0.13.2
shapely==2.0.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.23
sqlparse==0.5.0
tenacity==8.2.3
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
virtualenv==20.24.5
Werkzeug==2.3.7
wsproto==1.2.0
